+++ wildfly.py
@@ -12,7 +12,7 @@
 needed by more than one WildFly/JBoss plugin."""
 
 __author__ = 'Linuxfabrik GmbH, Zurich/Switzerland'
-__version__ = '2023112901'
+__version__ = '2024031501'
 
 from . import base
 from . import url
@@ -25,11 +25,20 @@ def get_data(args, data, uri=''):
         uri = '{}/host/{}/server/{}'.format(uri, args.NODE, args.INSTANCE)
     header = {'Content-Type': 'application/json'}
     result = base.coe(url.fetch_json(
-        uri, timeout=args.TIMEOUT,
-        header=header, data=data,
-        digest_auth_user=args.USERNAME, digest_auth_password=args.PASSWORD,
-        encoding='serialized-json'
+        uri,
+        data=data,
+        digest_auth_password=args.PASSWORD,
+        digest_auth_user=args.USERNAME,
+        encoding='serialized-json',
+        header=header,
+        insecure=args.INSECURE,
+        no_proxy=args.NO_PROXY,
+        timeout=args.TIMEOUT,
     ))
     if result['outcome'] != 'success':
-        base.oao('Error fetching data: "{}"'.format(result), STATE_UNKNOWN, always_ok=args.ALWAYS_OK)
+        base.oao(
+            'Error fetching data: "{}"'.format(result),
+            STATE_UNKNOWN,
+            always_ok=args.ALWAYS_OK,
+        )
     return result['result']